i18n: Indirect string-array declarations through arrays.xml (#1051)
* i18n: Indirect string-array declarations through arrays.xml.

Introduce string-array elements matching those from the source
strings.xml, but instead pointing to a @string reference.  The latter
is to be translated based on the assigned sub key.

Weblate does not handle string-arrays, but needs this indirection.

* Replace string-arrays with indirected translation source strings.

Move over the existing translations within string-arrays elements to
the string keys referenced in arrays.xml.

Only for the English (source) locale for now.

* Replace string-arrays with indirected translation strings.

Move over the existing translations within string-arrays elements to
the string keys referenced in arrays.xml.

For every language except the English (source) locale.