chore(strings): shorten crash notification text (#1337)
* chore(strings): shorten notification text (fixes #1113)

As suggested by @Catfriend1 in a
[comment](https://github.com/Catfriend1/syncthing-android/issues/1113#issuecomment-2553284015), we can just remove the
"Syncthing is/was..." text from the notification to reduce the length.

I've tested Sleeping and Running, but I don't know how to trigger the
Terminated condition.

* chore(strings): shorten crash notification text

This was one of the notification texts I glossed over in
Catfriend1#1319, I noticed it recently (ref Catfriend1#1336).